Rewrote templated recorder activity.
This version gets rid of the in-memory cast media lists and
replaces them with writes to the database. This will help ensure
that casts don't get lost while they're being created, even
if they're interrupted. It also improves the UI sync state,
as most UI widgets are now be data driven.

Removed the SegmentedProgressBar in favour of one backed by a
cursor.