Added network state detection to Sync.
Sync now checks to see if the system thinks it has a network connection
before syncing, to avoid errors. Will also stop a sync if the network
connection is lost.

Added intent extra for sync when the sync task was explicitly requested
by the user, so they can be notified if it failed due to a network
failure.

Fixed sync stopping.