Add ItineraryDetail and ItineraryList
Adds an itinerary list and some more tie-ins to the detail view. This
uses the new CursorLoader interface which should hopefully be a better
solution to loading data.

Updates the graphics to use appropriate icons. Layouts are adjusted
accordingly.

Twiddles the themes to break theme from the style. Attempts to unify the
coloring.

Modify the MediaProvider to simplify using Casts as child objects. First
steps to change this to a many-to-many relation are in place.