Improves nearby activity
Pulls in android support classes and modifies them to inherit from
MapActivity. (thanks to Carter for the suggestion!)
This makes all the unclosed cursor exceptions go away by allowing
LoaderManagers to be used properly.

Adds smarts to the Nearby activity so that it only refreshes from the
network after a timeout and only does local DB queries if the location
actually changed.

Abstracts the CastOverlay so that it can be used with locatable things other
than casts (events).

Sync now strips out any query strings when syncing children.