Adds event detail; prioritizes sync queue; more.
Adds EventDetail Activity.

Adds footer icon on all lists.

Fixes crash on LocatableItemOverlay; adds shadow back.

Improves handling on LocatableListWithMap.

Extracts some map utilities to a separate package.

Sync engine now uses a priority queue instead of a simple queue, so that
the most recent items get processed first. A queue was used instead of a
stack so that further prioritization can be done later down the road.

Adds sync request timeout so that items that are stale don't bunch up.

Adds a check to see if an item was sync'd recently (for both data and
media), to avoid pounding the server with requests.

Adds support for favorited.

First attempt to fix MapView bugs by putting adjacent views in separate
processes. Ugly and resource intensive, but seems to fix it on the test
devices.