Improves SyncEngine by storing the server's modified date
This should enable the system to be able to detect update conflicts as
well as function more reliably when the mobile's clock isn't stable.

The sync engine first checks the to see if the modified dates are
exactly the same. If they are, everything is clean. Otherwise, it
attempts to resolve version newness by comparing relative clocks and
entry edit ages.