Majorly overhauls the NetworkClient and account system
Instead of storing the base URL in a preferences file, this now saves it
in the Account metadata. This will make it work better when switching
between development and production accounts. This paves the way for
allowing multiple accounts on the same device (although there are still
many assumptions about how many accounts there are and such).

Removes some broken code from the NetworkClient that tried to show things
to the user.

This includes migration code to ensure that even existing accounts will
function properly if there is no base URL info stored in the account
metadata.