Moves DB helping code to SimpleContentProvider; NOTE: deletes DBs
This provides a much-needed move to using the SimpleContentProvider
to handle database-related storage. While this doesn't use most of
the features of the SimpleContentProvider, it paves the way to switching
over by swapping out the included helper libraries and pulling in
SimpleContentProvider as a dependency.

This includes a jar with a snapshot of SimpleContentProvider at
989192590d06c927ec9eb86f75eb718ccbff3c77 .