Rips out UI-related part of Locast to make LocastCore
In order to make the code more reusable, LocastCore will provide the
data, account, and sync layer for Locast as a library project, while
OpenLocast will be an example of using LocastCore.

This is a first, fairly brutal pass at removing unneeded UI parts. Some
files are modified in order to remove dependencies on UI features.