Brutally hacks apart Locast Core from its GUI; rough first pass
This is the first barrage of changes, intended to clear the way for a
separation of Locast Core from it GUI. This is most certainly incomplete
and also broken, but is also a good first pass. This requires a recent
version of SimpleContentProvider which will be included in subsequent
changes.