Adds support for updating modified date
SimpleContentProvider now has an auto_now which will automatically
update a field when the row is updated. This adds that to the modified
date for JsonSyncableItems.

While very convenient, this may need some reworking of the SyncEngine to
take this date into account as any updates (from the server, not just
the client) will result in a sync.