Reverts the use of AUTO_NOW for modified date
AUTO_NOW kept triggering the client to PUT its local data to the server
whenever any of the columns were updated by anything, including the
sync engine.

This adds a special parameter that's sent to the content provider to
disable auto-updating of the modified date upon update()s. The sync
engine sends that with all its requests.