Major rework of the SyncEngine
DB updates: adds "dirty" column to JsonSyncableItem

This changes the way the sync engine checks to see if content is up to
date by introducing a "dirty" column. Whenever an insert or update is
done on the database using the SyncableSimpleContentProvider, it marks
that row dirty (unless SyncableProvider.FLAG_DO_NOT_CHANGE_DIRTY is
used). Dirty rows will be published up to the server. To go the other
direction, the last modified date sent from the server is sent and
compared. This should hopefully fix the issue that age-based
synchronization was having (lots more GET/PUT churn than expected).

It now only does age-based synchronization when both sides have been
marked dirty. While not the best conflict resolution, it's a start and
makes sense for resolving conflicts done by the same user.