Automatically filter out deleted items
Currently, deletion is a two-step process. Items are marked for deletion
by setting a flag. Once marked for deletion, the SyncEngine will attempt
to propagate the deletions to the server. Once confirmed deleted from
the server, they will be properly deleted.

This makes this even more transparent by automatically filtering out any
deleted items unless a special query is made. This makes marking an item
for deletion approximately the same as deleting it - as far as queries
go. A special flag can be added to return all items, including deleted
ones, which the SyncEngine uses.