Requires app to implement interface to hook in NetworkClient
In order to make dependency injection possible for NetworkClient, this
adds a requirement for any apps wishing to use this library.

As there's no way beyond extending the Application object to inject
resources into a context, this does just that. Now any app wishing to
use this library must have its Application implement the
LocastApplicationCallbacks interface. There's a convenient
LocastApplication class which extends Application and implements this
callback.