Adds a mandatory UUID field to JsonSyncableItem
This introduces a protocol change.

In order to help prevent duplicates, this adds a new UUID field that
should be populated at the time of creating a new content item. If there
is network trouble between creation and posting to acquire a public URL,
a duplicate would sometimes be created. This prevents that scenario from
occurring by detecting conflicts when posting content with an already
known UUID. The sync engine can automatically resolve these conflicts
and update the client with the known information.